# Fix log plugin benchmark and slightly improve performance