--- a/plugin/log/log.go
+++ b/plugin/log/log.go
@@ -37,13 +37,13 @@ func (l Logger) ServeDNS(ctx context.Context, w dns.ResponseWriter, r *dns.Msg) 
 
 		// If we don't set up a class in config, the default "all" will be added
 		// and we shouldn't have an empty rule.Class.
-		_, ok := rule.Class[response.All]
-		if !ok {
+		_, shouldLog := rule.Class[response.All]
+		if !shouldLog {
 			tpe, _ := response.Typify(rrw.Msg, time.Now().UTC())
 			class := response.Classify(tpe)
-			_, ok = rule.Class[class]
+			_, shouldLog = rule.Class[class]
 		}
-		if ok {
+		if shouldLog {
 			logstr := l.repl.Replace(ctx, state, rrw, rule.Format)
 			clog.Infof(logstr)
 		}