--- a/plugin/log/log_test.go
+++ b/plugin/log/log_test.go
@@ -3,6 +3,7 @@ package log
 import (
 	"bytes"
 	"context"
+	"io/ioutil"
 	"log"
 	"strings"
 	"testing"
@@ -239,8 +240,7 @@ func TestLogged(t *testing.T) {
 }
 
 func BenchmarkLogged(b *testing.B) {
-	var f bytes.Buffer
-	log.SetOutput(&f)
+	log.SetOutput(ioutil.Discard)
 
 	rule := Rule{
 		NameScope: ".",