### 1. Why is this pull request needed and what does it do?

It fixes memory/alloc reporting for `BenchmarkLogged` and provides a small improvement to the logging logic.

Benchmark results:
```
benchmark              old ns/op     new ns/op     delta
BenchmarkLogged-12     5742          5573          -2.94%

benchmark              old allocs     new allocs     delta
BenchmarkLogged-12     60             60             +0.00%

benchmark              old bytes     new bytes     delta
BenchmarkLogged-12     4536          4145          -8.62%
```

### 2. Which issues (if any) are related?

N/A

### 3. Which documentation changes (if any) need to be made?

N/A

### 4. Does this introduce a backward incompatible change or deprecation?


N/A