@miekg 2019-07-14
[ Quoting <notifications@github.com> in "Re: [coredns/coredns] Fix log plugi..." ]
>-		_, ok1 := rule.Class[class]
>-		if ok || ok1 {
>+		if !ok {
>
>The variable name isn't great, but the goal here is to skip the call to [`response.Typify()`](https://github.com/charlievieth/coredns/blob/cev/log-perf/plugin/pkg/response/typify.go#L57) if we already know the message needs to be logged: https://github.com/coredns/coredns/blob/72c84dcb1cb7e69941eca7050153e80a5c9f1cf7/plugin/log/log.go#L42

sure, but I had to read this 3 times to make sure it did the right thing. This 
doesn't sound super critical for performance, so I rather have it more readable.