@charlievieth 2019-07-15
Thanks again for the quick feedback - please let me know if there are any other changes you'd like for me to make.